feat: add output folder to workspace
This commit will allow user to choose a quick pick option on transform and customizations to choose to add the output folder to workspace. For now, the output is stored in /home/swaingotnochill/.m2kplugi
noutput/output directory. If a directory by name output is already present, the name is further modified as out - timestamp. This allows to store multiple outputs without bothering about overwrite flag or manually deleting it.

Signed-off-by: Roshan Swain <swainroshan001@gmail.com>

workinporgress

Signed-off-by: Roshan Swain <swainroshan001@gmail.com>

fix: keeps the outputfolder in the same level as source

This commit removes the previous implemenation of generating .m2kpluginoutput in home directory. Instead it creates a m2kpluginoutput in the same level as the src folder in which the command was clicked.

Signed-off-by: Roshan Swain <swainroshan001@gmail.com>

clean: removing out folder as mistakenly pushed

clean: update .gitignore to ignore out directory

Signed-off-by: Roshan Swain <swainroshan001@gmail.com>

clean: removing .vsix package file and adding it to the .gitignore

Signed-off-by: Roshan Swain <swainroshan001@gmail.com>